+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 2026-02-28
+### Changed
+- Made docker content scrollable using QScrollArea
+
 ## 2026-02-07
 ### Changed
 - Floating widget visibility (Tool Options, Color Selector) now auto-persists to config on toggle